+++ .idea/compiler.xml
@@ -7,10 +7,12 @@
         <processorPath useClasspath="false">
           <entry name="$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.projectlombok/lombok/1.18.36/5a30490a6e14977d97d9c73c924c1f1b5311ea95/lombok-1.18.36.jar" />
         </processorPath>
-        <module name="ecommerce-msa-server.main" />
+        <module name="ecommerce-msa-server.jwt.main" />
         <module name="ecommerce-msa-server.product-service.main" />
-        <module name="ecommerce-msa-server.order-service.main" />
         <module name="ecommerce-msa-server.user-service.main" />
+        <module name="ecommerce-msa-server.gateway-server.main" />
+        <module name="ecommerce-msa-server.order-service.main" />
+        <module name="ecommerce-msa-server.payment-service.main" />
         <module name="ecommerce-msa-server.common.main" />
       </profile>
     </annotationProcessing>
