+++ docker-compose.yml
@@ -75,7 +75,6 @@ services:
       - SPRING_DATASOURCE_PASSWORD=mysqlpw
       - SPRING_JPA_PROPERTIES_HIBERNATE_DIALECT=org.hibernate.dialect.MySQL8Dialect
       - SPRING_JPA_HIBERNATE_DDL_AUTO=create-drop
-      - SPRING_JWT_SECRET=dudguswntjdtkdwlsdbstjrahenahencnldjqtjdrhd
       - EUREKA_SERVER_HOST=eureka-server
 
 
@@ -93,7 +92,6 @@ services:
       - SPRING_DATASOURCE_PASSWORD=mysqlpw
       - SPRING_JPA_PROPERTIES_HIBERNATE_DIALECT=org.hibernate.dialect.MySQL8Dialect
       - SPRING_JPA_HIBERNATE_DDL_AUTO=create-drop
-      - SPRING_JWT_SECRET=dudguswntjdtkdwlsdbstjrahenahencnldjqtjdrhd
       - EUREKA_SERVER_HOST=eureka-server
 networks:
   msa_network:
