+++ gateway-server/build.gradle
@@ -39,6 +39,7 @@ dependencies {
     implementation 'io.jsonwebtoken:jjwt-jackson:0.12.3'
 
     implementation project(':common')
+    implementation project(':jwt')
 }
 
 ext {
