+++ gateway-server/src/main/java/com/msa/gateway/GatewayServerApplication.java
@@ -4,7 +4,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
 
-@ComponentScan(basePackages = {"com.msa.gateway", "com.msa.common"})
+@ComponentScan(basePackages = {"com.msa.gateway", "com.msa.common", "com.msa.jwt"})
 @SpringBootApplication
 public class GatewayServerApplication {
     public static void main(String[] args) {
