+++ settings.gradle
@@ -6,3 +6,5 @@ include 'product-service'
 include 'payment-service'
 include 'eureka-server'
 include 'gateway-server'
+include 'jwt'
+
