+++ user-service/build.gradle
@@ -50,6 +50,7 @@ dependencies {
     implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
 
     implementation project(':common')
+    implementation project(':jwt')
 }
 
 ext {
