+++ user-service/src/main/java/com/msa/user/UserServiceApplication.java
@@ -4,7 +4,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
 
-@ComponentScan(basePackages = {"com.msa.common", "com.msa.user"})
+@ComponentScan(basePackages = {"com.msa.jwt", "com.msa.common", "com.msa.user"})
 @SpringBootApplication
 public class UserServiceApplication {
     public static void main(String[] args) {
