+++ user-service/src/test/java/com/msa/user/adapter/in/web/filter/AuthenticationFilterTest.java
@@ -8,7 +8,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.msa.common.utils.JwtUtil;
+import com.msa.jwt.JwtUtil;
 import com.msa.user.adapter.in.web.dto.request.UserLoginRequest;
 import com.msa.user.application.service.CustomUserDetailsService;
 import jakarta.servlet.FilterChain;
@@ -22,12 +22,10 @@ import com.msa.user.exception.InvalidLoginException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
